+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.8.0] - 2020-03-25
+
+### Added
+- ability to send message to multiple receivers
+
+### Changed
+- Refactored all Screens to CUBA 7 APIs
+
+### Dependencies
+- CUBA 7.2.x
+
 ## [0.7.0] - 2020-03-21
 
 ### Added